FIT-81-Domingo 30

+++ src/main/java/com/fitbum/entidades/estadisticas/Estadisticas.java
@@ -22,8 +22,8 @@ public class Estadisticas {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-
-    private Integer	idEstadisticas;
+// Ale: He cambiado Integer por Long
+    private Long idEstadisticas;
     private String	descripcion;
     @JsonBackReference
     @ManyToOne(fetch = FetchType.LAZY)
